Moved AbstractTest classes to pytest. Added sample-changer-mockup.xml
