+++ test/pytest/TestAbstractMotorBase.py
@@ -28,7 +28,7 @@ __date__ = "09/04/2020"
 
 import abc
 import pytest
-from HardwareRepository.HardwareObjects.abstract.testing import TestAbstractActuatorBase
+from HardwareRepository.test.pytest import TestAbstractActuatorBase
 
 test_object = TestAbstractActuatorBase.test_object
