+++ test/pytest/test_detector.py
@@ -22,7 +22,7 @@ from __future__ import division, absolute_import
 from __future__ import print_function, unicode_literals
 
 import pytest
-from HardwareRepository.HardwareObjects.abstract.testing import TestHardwareObjectBase
+from HardwareRepository.test.pytest import TestHardwareObjectBase
 
 __copyright__ = """ Copyright © 2016 - 2020 by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
