+++ test/pytest/test_detector_distance.py
@@ -22,7 +22,7 @@ from __future__ import division, absolute_import
 from __future__ import print_function, unicode_literals
 
 import pytest
-from HardwareRepository.HardwareObjects.abstract.testing import TestAbstractMotorBase
+from HardwareRepository.test.pytest import TestAbstractMotorBase
 
 __copyright__ = """ Copyright © 2016 - 2020 by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
